Updated Unit tests for FeatureTypeStyle and Rule.
--HG--
branch : devel-0.2
rename : src/test/java/org/geolatte/maprenderer/sld/SLDPainterTest.java => src/test/java/org/geolatte/maprenderer/sld/BaseFeatureTypeStyleTest.java
rename : src/test/java/org/geolatte/maprenderer/sld/TestFeatureTypeStylePainter.java => src/test/java/org/geolatte/maprenderer/sld/TestFeatureTypeStyleRule.java

+++ src/main/java/org/geolatte/maprenderer/java2D/JAIMapGraphics.java
@@ -42,7 +42,7 @@ public class JAIMapGraphics extends MapGraphics {
 
     private final static Color DEFAULT_BACKGROUND_COLOR = new Color(1.0f, 1.0f, 1.0f, 0.0f);
 
-    private float scale;
+    private double scale;
     private final int width;
     private final int height;
     private Graphics2D g2;
@@ -134,8 +134,8 @@ public class JAIMapGraphics extends MapGraphics {
         AffineTransform atf = new AffineTransform();
         double sx = width / extent.getWidth();
         double sy = height / extent.getHeight();
-        this.scale = (float) Math.min(sx, sy);
-        //TODO -- shouldn't I maintain aspect-ratio?
+        this.scale = Math.min(sx, sy);
+        //TODO -- shouldn't we maintain aspect-ratio?
         atf.scale(sx, -sy);
         atf.translate(-extent.getMinX(), -extent.getMaxY());
         setTransform(atf);
@@ -152,7 +152,7 @@ public class JAIMapGraphics extends MapGraphics {
     }
 
     @Override
-    public float getScale() {
+    public double getScale() {
         return this.scale;
     }
