+++ src/main/java/org/geolatte/maprenderer/map/MapGraphics.java
@@ -26,7 +26,15 @@ import org.geolatte.maprenderer.reference.SpatialReference;
 import java.awt.*;
 import java.awt.image.RenderedImage;
 
+/**
+ * A specialized <code>Graphics2D</code> for rendering spatial data.
+ */
 public abstract class MapGraphics extends Graphics2D {
+
+//    TODO : change SpatialReference to geolatte-geom's CRS
+//    TODO : remove setToExtent() and have extent set in constructor
+//    TODO : add method to set a DPI "hint" == intended resolution
+//    TODO : getScale() should provide reasonable value in case of geodetic CRS's. (see symbology encoding standard 05-077r4, p. 10)
 
     abstract public Dimension getDimension();
 
@@ -36,6 +44,6 @@ public abstract class MapGraphics extends Graphics2D {
 
     abstract public void setToExtent(SpatialExtent extent);
 
-    abstract public float getScale();
+    abstract public double getScale();
 
 }
